backend/pyproject.toml
[project]
name = "whisper-hebrew-trainer"
version = "0.1.0"
description = "Hebrew speech training app with Whisper fine-tuning"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.12",
    "faster-whisper>=1.1.0",
    "jiwer>=3.0.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.27.0",
    "google-generativeai>=0.8.0",
    "google-cloud-aiplatform>=1.40.0",
]

[project.optional-dependencies]
training = [
    "transformers>=4.44.0",
    "peft>=0.13.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "torch>=2.4.0",
    "evaluate>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
include = ["app*", "training*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
